updated api package structure,- extracted MainRepository into AthletesRepositoy and FeedsRepository
