+++ .idea/deploymentTargetDropDown.xml
@@ -12,6 +12,6 @@
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-01-04T23:53:00.052097600Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-01-05T18:23:35.711951600Z" />
   </component>
 </project>
