+++ .idea/misc.xml
@@ -7,7 +7,7 @@
         <entry key="..\:/Users/david/OneDrive/Desktop/AsyncLabs/app/src/main/res/layout/activity_splash.xml" value="0.25729166666666664" />
         <entry key="..\:/Users/david/OneDrive/Desktop/AsyncLabs/app/src/main/res/layout/activity_splash2.xml" value="0.15555555555555556" />
         <entry key="..\:/Users/david/OneDrive/Desktop/AsyncLabs/app/src/main/res/layout/athlete_feed_item.xml" value="0.1" />
-        <entry key="..\:/Users/david/OneDrive/Desktop/AsyncLabs/app/src/main/res/layout/fragment_athletes.xml" value="0.24739583333333334" />
+        <entry key="..\:/Users/david/OneDrive/Desktop/AsyncLabs/app/src/main/res/layout/fragment_athletes.xml" value="0.1" />
         <entry key="..\:/Users/david/OneDrive/Desktop/AsyncLabs/app/src/main/res/layout/fragment_home.xml" value="0.1" />
         <entry key="..\:/Users/david/OneDrive/Desktop/AsyncLabs/app/src/main/res/layout/fragment_splash.xml" value="0.2713541666666667" />
         <entry key="..\:/Users/david/OneDrive/Desktop/AsyncLabs/app/src/main/res/layout/layout_video_list_item.xml" value="0.1" />
