+++ app/src/main/java/com/asynclabs/asyncsport/data/model/AthleteResponse.kt
@@ -1,4 +1,4 @@
-package com.asynclabs.asyncsport.api.model
+package com.asynclabs.asyncsport.data.model
 
 import com.google.gson.annotations.SerializedName
 import com.google.gson.annotations.Expose
