+++ app/src/main/java/com/asynclabs/asyncsport/data/model/FeedResponse.kt
@@ -1,4 +1,4 @@
-package com.asynclabs.asyncsport.api
+package com.asynclabs.asyncsport.data
 
 
 import com.google.gson.annotations.SerializedName
