+++ app/src/main/java/com/asynclabs/asyncsport/data/repository/AthletesRepository.kt
@@ -1,0 +1,7 @@
+package com.asynclabs.asyncsport.data.repository
+
+import com.asynclabs.asyncsport.data.model.AthleteResponse
+
+interface AthletesRepository {
+    suspend fun getAthletes():List<AthleteResponse>
+}
