+++ app/src/main/java/com/asynclabs/asyncsport/data/repository/FeedsRepository.kt
@@ -1,0 +1,7 @@
+package com.asynclabs.asyncsport.data.repository
+
+import com.asynclabs.asyncsport.data.FeedResponse
+
+interface FeedsRepository {
+    suspend fun getAllFeeds(page:Int, sport:String):List<FeedResponse>
+}
