--- app/src/main/java/com/asynclabs/asyncsport/model/FeedModel.kt
@@ -1,2 +1,0 @@
-package com.asynclabs.asyncsport.model
-
