+++ app/src/main/java/com/asynclabs/asyncsport/ui/athletes/AthletesViewModel.kt
@@ -3,14 +3,11 @@ package com.asynclabs.asyncsport.ui.athletes
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.asynclabs.asyncsport.api.model.AthleteResponse
+import com.asynclabs.asyncsport.data.model.AthleteResponse
 
-import com.asynclabs.asyncsport.repository.MainRepository
+import com.asynclabs.asyncsport.data.repository.MainRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 import javax.inject.Inject
 
 @HiltViewModel
