+++ app/src/main/java/com/asynclabs/asyncsport/ui/home/HomeViewModel.kt
@@ -3,13 +3,10 @@ package com.asynclabs.asyncsport.ui.home
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.asynclabs.asyncsport.api.FeedResponse
-import com.asynclabs.asyncsport.repository.MainRepository
+import com.asynclabs.asyncsport.data.FeedResponse
+import com.asynclabs.asyncsport.data.repository.MainRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 import javax.inject.Inject
 
 @HiltViewModel
